moving some files to the main module
