+++ src/main/java/espresso/ApiConfiguration.java
@@ -1,4 +1,4 @@
-package espresso.achievement.service;
+package espresso;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.support.ResourceBundleMessageSource;
