+++ src/main/java/espresso/ApiMessageHelper.java
@@ -1,4 +1,4 @@
-package espresso.achievement.service;
+package espresso;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.MessageSource;
