+++ src/main/java/espresso/achievement/application/queryHandlers/AchievementMediaQueryHandler.java
@@ -3,6 +3,7 @@ package espresso.achievement.application.queryHandlers;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import espresso.AppConfiguration;
 import espresso.achievement.application.response.HandlerResult;
 import espresso.achievement.domain.contracts.IAchievementMediaQueryHandler;
 import espresso.achievement.domain.contracts.IMediaQryRepository;
@@ -10,7 +11,6 @@ import espresso.achievement.domain.entities.KeyGenerator;
 import espresso.achievement.domain.entities.PreMedia;
 import espresso.achievement.domain.queries.GetAchievementMediaStorageQuery;
 import espresso.achievement.domain.readModels.MediaStorageDetailReadModel;
-import espresso.achievement.service.AppConfiguration;
 import lombok.NoArgsConstructor;
 
 @Service
