+++ src/main/java/espresso/achievement/service/AchievementQryApi.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import espresso.ApiMessageHelper;
 import espresso.achievement.application.response.HandlerResult;
 import espresso.achievement.domain.contracts.IAchievementMediaQueryHandler;
 import espresso.achievement.domain.contracts.IAchievementQueryHandler;
